# openapi.yaml
openapi: 3.0.3
info:
  title: WhatsApp API Server
  description: |
    REST API for WhatsApp automation using whatsapp-web.js with PM2 support.
    
    ## Quick Start (Hebrew)
    1. בדוק בריאות: GET /api/health
    2. התחבר לוואטסאפ: POST /api/connect  
    3. קבל QR: GET /api/qr
    4. שלח הודעה: POST /api/send
    
    ## Authentication
    Use the `x-api-key` header or `apiKey` query parameter for authentication.
    
  version: 1.0.0
  contact:
    name: AsafBen
    email: asaf@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
        hebrew:
          type: object
          properties:
            message:
              type: string
              description: Message in Hebrew
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        hebrew:
          type: object
          properties:
            error:
              type: string
              description: Error message in Hebrew
    WhatsAppStatus:
      type: object
      properties:
        isReady:
          type: boolean
          description: Whether WhatsApp client is ready
        connectionStatus:
          type: string
          description: Current connection status
          enum: [not_initialized, connecting, connected, authenticated, ready, disconnected]
        qrCode:
          type: string
          nullable: true
          description: QR code for authentication (if available)
        clientInfo:
          type: object
          nullable: true
          properties:
            number:
              type: string
              description: WhatsApp number
            name:
              type: string
              description: Profile name
    SendMessageRequest:
      type: object
      required:
        - phoneNumber
        - message
      properties:
        phoneNumber:
          type: string
          description: Phone number without country code
          example: "0501234567"
        message:
          type: string
          description: Message text to send
          example: "שלום! זוהי הודעת בדיקה"
          maxLength: 4096
        countryCode:
          type: string
          description: Country code (default is 972 for Israel)
          example: "972"
          default: "972"
...
paths:
...
  /api/status:
    get:
      summary: Get WhatsApp Status
      description: Get current WhatsApp connection status and client information
      tags:
        - WhatsApp Connection
      security: []
      responses:
        '200':
          description: WhatsApp status information
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      status:
                        allOf:
                          - $ref: '#/components/schemas/WhatsAppStatus'
                          - type: object
                            properties:
                              lastUpdate:
                                type: string
                                format: date-time
        '500':
          description: Failed to get status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/connect:
    post:
      summary: Connect to WhatsApp
      description: Initialize WhatsApp connection and start authentication process
      tags:
        - WhatsApp Connection
      responses:
        '200':
          description: Connection initiated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      nextSteps:
                        type: array
                        items:
                          type: string
                        description: Next steps to complete setup
        '500':
          description: Failed to connect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/qr:
    get:
      summary: Get QR Code
      description: Get QR code for WhatsApp authentication
      tags:
        - WhatsApp Connection  
      security: []
      responses:
        '200':
          description: QR code information
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      qrCode:
                        type: string
                        nullable: true
                        description: Base64 encoded QR code image
        '400':
          description: WhatsApp service not initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to get QR code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/send:
    post:
      summary: Send Message
      description: Send a text message to a WhatsApp number
      tags:
        - Messaging
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            examples:
              hebrew_message:
                summary: Hebrew message example
                value:
                  phoneNumber: "0501234567"
                  message: "שלום! איך השלום?"
                  countryCode: "972"
              english_message:
                summary: English message example  
                value:
                  phoneNumber: "0501234567"
                  message: "Hello! How are you?"
                  countryCode: "972"
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      messageId:
                        type: string
                        description: Unique message identifier
                      chatId:
                        type: string
                        description: Chat identifier
        '400':
          description: Invalid request or WhatsApp not ready
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - type: object
                    properties:
                      errors:
                        type: array
                        items:
                          type: string
                        description: Validation errors
                      help:
                        type: string
                        description: Help message
        '500':
          description: Failed to send message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...